Set specifikation

+++ app/Http/Controllers/TodosController.php
@@ -14,7 +14,7 @@ class TodosController extends Controller
    */
   public function index()
   {
-      return Todo::all();
+      return Todo::where('user_id', auth()->user()->id)->get();
   }
 
   /**
@@ -30,7 +30,11 @@ class TodosController extends Controller
           'completed' => 'required|boolean'
       ]);
 
-      $todo = Todo::create($data);
+      $todo = Todo::create([
+        'user_id' => auth()->user()->id,
+        'title'   => $request->title,
+        'completed' =>$request->completed
+      ]);
 
       return response($todo, 201);
   }
@@ -44,6 +48,16 @@ class TodosController extends Controller
    */
   public function update(Request $request, Todo $todo)
   {
+
+    /**
+     * Karna sudah ada $todo jadi otomatis sudah mendapatkan
+     * id todo mana yang harus diupdate
+     */
+
+    if ($todo->user_id !== auth()->user()->id) {
+      return response()->json('Unauthorized', 401);
+    }
+
     $data = $request->validate([
         'title'   => 'required|string',
         'completed' => 'required|boolean'
@@ -60,7 +74,7 @@ class TodosController extends Controller
       'completed' => 'required|boolean'
     ]);
 
-    Todo::query()->update($data); // update all todos completed to true
+    Todo::where('user_id', auth()->user()->id)->update($data); // update completed todos menjadi true berdasarkan user_id
 
     return response('Updated all completed', 200);
   }
@@ -72,6 +86,9 @@ class TodosController extends Controller
    * @return \Illuminate\Http\Response
    */
   public function destroy(Todo $todo) {
+    if ($todo->user_id !== auth()->user()->id) {
+      return response()->json('Unauthorized', 401);
+    }
 
     $todo->delete();
     return response('Deleted todo item', 200);
@@ -80,13 +97,31 @@ class TodosController extends Controller
 
   public function destroyCompleted(Request $request) {
 
+    // ambil id yang mau di delete
+    $todosToDelete = $request->todos;
+    // ambil id todo berdasarkan auth user
+    $userTodoIds = auth()->user()->todos->map(function($todo){
+      return $todo->id;
+    });
+
+    /**
+     * The every method may be used to verify that all elements of a collection pass a given truth test:
+     */
+    // apakah $userTodoIds berisi $todoToDelete
+    $valid = collect($todosToDelete)->every(function($value, $key) use ($userTodoIds) {
+      return $userTodoIds->contains($value);
+    });
+    // Lakukan validasi
+    if(!$valid){
+      return response()->json('Unauthorized', 401); 
+    }
+
     $request->validate([
       'todos'   => 'required|array',
     ]);
-
+    // Delete
     Todo::destroy($request->todos); // deleted sekaligus
 
-    return response('Deleted', 200);
-
+    return response()->json('Deleted', 200);
   }
 }
